+++ Sources/MereRunCLI/MereRunCLI.swift
@@ -25,6 +25,7 @@ struct MereRunCLI: AsyncParsableCommand {
             Music.self,
             Video.self,
             Model.self,
+            Status.self,
             API.self,
             Setup.self,
             Agent.self,